+++ cmd/encipher.go
@@ -2,12 +2,19 @@ package cmd
 
 import (
 	"fmt"
+	"os"
+	"strings"
 
 	"github.com/ccaroon/enigmachine/enigma"
 	"github.com/spf13/cobra"
 )
 
-var keyFlag string
+var (
+	keyFlag           string
+	keepOrigFmtFlag   bool
+	blockSizeFlag     int
+	blocksPerLineFlag int
+)
 
 var encodeCmd = &cobra.Command{
 	Use:     "encode <message>",
@@ -15,9 +22,12 @@ var encodeCmd = &cobra.Command{
 	Short:   "Encode a message",
 	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
+		var content string
 		input := args[0]
 
-		key := enigma.ParseKeySpec(keyFlag)
+		key, err := enigma.ParseKeySpec(keyFlag)
+		handleCmdError(err)
+
 		machine := enigma.NewEnigma(
 			key.ReflId,
 			key.RotorIds,
@@ -25,13 +35,29 @@ var encodeCmd = &cobra.Command{
 		)
 		machine.ConfigureRotors(key.RotorCfg)
 
-		output := machine.EncipherString(input)
+		if strings.HasPrefix(input, "@") {
+			data, err := os.ReadFile(input[1:])
+			handleCmdError(err)
+
+			content = string(data)
+		} else {
+			content = input
+		}
+
+		options := enigma.NewOutputOptions(blockSizeFlag, blocksPerLineFlag, keepOrigFmtFlag)
+		output := machine.EncipherString(content, options)
 		fmt.Println(output)
 	},
 }
 
 func init() {
 	encodeCmd.Flags().StringVarP(&keyFlag, "key", "k", "B:I,II,III:", "Key/Configuration to use to encode the message. E.g. B:I@F,II,II@X:AZ,QR,XM")
+
+	encodeCmd.Flags().BoolVarP(&keepOrigFmtFlag, "original", "o", false, "Keep original input formatting, including punctuation, spacing & lines.")
+
+	encodeCmd.Flags().IntVarP(&blockSizeFlag, "block-size", "b", 5, "Group encoded letter into blocks of this size.")
+
+	encodeCmd.Flags().IntVarP(&blocksPerLineFlag, "line-size", "l", 15, "The number of blocks per line.")
 
 	rootCmd.AddCommand(encodeCmd)
 }